Refactor registers.rs from debug/mod.rs
